personal profile under learner to read the id

+++ Course station/Course station/Controllers/PersonalProfileController.cs
@@ -6,6 +6,7 @@ using Microsoft.EntityFrameworkCore;
 
 namespace Course_station.Controllers
 {
+    [Route("Learners/{learnerId}/[controller]")]
     public class PersonalProfileController : Controller
     {
         private readonly ApplicationDbContext _context;
@@ -15,33 +16,44 @@ namespace Course_station.Controllers
             _context = context;
         }
 
-        // GET: PersonalProfile
-        public async Task<IActionResult> Index()
+        // GET: Learners/{learnerId}/PersonalProfile
+        [HttpGet]
+        public async Task<IActionResult> Index(int learnerId)
         {
-            return View(await _context.PersonalProfiles.ToListAsync());
+            var personalProfiles = await _context.PersonalProfiles
+                .Where(p => p.LearnerId == learnerId)
+                .ToListAsync();
+            ViewBag.LearnerId = learnerId;
+            return View(personalProfiles);
         }
 
-        // GET: PersonalProfile/Create
-        public IActionResult Create()
+        // GET: Learners/{learnerId}/PersonalProfile/Create
+        [HttpGet("Create")]
+        public IActionResult Create(int learnerId)
         {
+            ViewBag.LearnerId = learnerId;
             return View();
         }
 
-        // POST: PersonalProfile/Create
-        [HttpPost]
+        // POST: Learners/{learnerId}/PersonalProfile/Create
+        [HttpPost("Create")]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("LearnerId,ProfileId,PreferedContentType,EmotionalState,PersonalityType")] PersonalProfile personalProfile)
+        public async Task<IActionResult> Create(int learnerId, [Bind("ProfileId,PreferedContentType,EmotionalState,PersonalityType")] PersonalProfile personalProfile)
         {
             if (ModelState.IsValid)
             {
+                personalProfile.LearnerId = learnerId;
                 _context.Add(personalProfile);
                 await _context.SaveChangesAsync();
-                return RedirectToAction(nameof(Index));
+                return RedirectToAction(nameof(Index), new { learnerId });
             }
+            ViewBag.LearnerId = learnerId;
             return View(personalProfile);
         }
-        // GET: PersonalProfile/Delete/5
-        public async Task<IActionResult> Delete(int? id)
+
+        // GET: Learners/{learnerId}/PersonalProfile/Delete/5
+        [HttpGet("Delete/{id}")]
+        public async Task<IActionResult> Delete(int learnerId, int? id)
         {
             if (id == null)
             {
@@ -49,24 +61,29 @@ namespace Course_station.Controllers
             }
 
             var personalProfile = await _context.PersonalProfiles
-                .FirstOrDefaultAsync(m => m.ProfileId == id);
+                .FirstOrDefaultAsync(m => m.ProfileId == id && m.LearnerId == learnerId);
             if (personalProfile == null)
             {
                 return NotFound();
             }
 
+            ViewBag.LearnerId = learnerId;
             return View(personalProfile);
         }
 
-        // POST: PersonalProfile/Delete/5
-        [HttpPost, ActionName("Delete")]
+        // POST: Learners/{learnerId}/PersonalProfile/Delete/5
+        [HttpPost("Delete/{id}"), ActionName("Delete")]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> DeleteConfirmed(int id)
+        public async Task<IActionResult> DeleteConfirmed(int learnerId, int id)
         {
-            var personalProfile = await _context.PersonalProfiles.FindAsync(id);
-            _context.PersonalProfiles.Remove(personalProfile);
-            await _context.SaveChangesAsync();
-            return RedirectToAction(nameof(Index));
+            var personalProfile = await _context.PersonalProfiles
+                .FirstOrDefaultAsync(m => m.ProfileId == id && m.LearnerId == learnerId);
+            if (personalProfile != null)
+            {
+                _context.PersonalProfiles.Remove(personalProfile);
+                await _context.SaveChangesAsync();
+            }
+            return RedirectToAction(nameof(Index), new { learnerId });
         }
     }
 }
